initial creation of shared tester package
