+++ experiments/tester/.npmignore
@@ -1,0 +1,13 @@
+src
+node_modules
+.gitignore
+.gitattributes
+.editorconfig
+config.js
+jest.config.js
+tslint.json
+tsconfig.json
+jsconfig.json
+webpack.config.js
+webpack.serve.config.js
+*.build.log
