+++ experiments/tester/babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  presets: ['module:metro-react-native-babel-preset'],
+  babelrcRoots: ['experiments/*']
+};
