+++ experiments/tester/jest.setup.js
@@ -1,0 +1,6 @@
+// @ts-check
+
+const enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+
+enzyme.configure({ adapter: new Adapter() });
