+++ experiments/tester/just.config.js
@@ -1,0 +1,3 @@
+const { preset } = require('@uifabricshared/build-native');
+
+preset();
