+++ experiments/tester/react-native.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  reactNativePath: '../../node_modules/@office-iss/react-native-win32'
+};
