+++ experiments/tester/src/RNTester/TestComponents/Square.tsx
