+++ experiments/tester/tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "target": "es5",
+    "jsx": "react",
+    "outDir": "lib",
+    "declaration": true,
+    "sourceMap": true,
+    "experimentalDecorators": true,
+    "importHelpers": true,
+    "strictNullChecks": true,
+    "noImplicitAny": true,
+    "noEmitOnError": true,
+    "skipLibCheck": true,
+    "moduleResolution": "node",
+    "noUnusedLocals": true,
+    "strict": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "lib": ["es6"],
+    "types": ["node", "webpack-env", "jest"]
+  },
+  "include": ["src"]
+}
